requirements.txt
pytest
cloudpickle
cryptography
mega.py
